Conflicts:
       contrib/cmake/CMakeLists.txt
       src/MEGASync/MegaApplication.cpp
       src/MEGASync/MegaApplication.h
       src/MEGASync/gui/InfoDialog.h
       src/MEGASync/gui/Resources_linux.qrc
       src/MEGASync/gui/Resources_macx.qrc
       src/MEGASync/gui/Resources_win.qrc
       src/MEGASync/gui/SettingsDialog.cpp
       src/MEGASync/gui/SettingsDialog.h
       src/MEGASync/gui/gui.pri
       src/MEGASync/gui/images/settings-contacts.png
       src/MEGASync/gui/images/settings-contacts@2x.png
       src/MEGASync/gui/linux/SettingsDialog.ui
       src/MEGASync/gui/win/SettingsDialog.ui

+++ src/MEGASync/gui/CustomTransferItem.cpp
@@ -365,7 +365,7 @@ void CustomTransferItem::updateTransfer()
     }
 
     // Update progress bar
-    unsigned int permil = (totalSize > 0) ? ((1000 * totalTransferredBytes) / totalSize) : 0;
+    unsigned int permil = (totalSize > 0) ? static_cast<unsigned int>((1000 * totalTransferredBytes) / totalSize) : 0;
     ui->pbTransfer->setValue(permil);
 }
